Split crypto crate into hasher and signer crates
