+++ README.md
@@ -52,7 +52,7 @@ let verifier_index = prover_index.verifier_index();
 
 // create a proof
 let group_map = <Affine as CommitmentCurve>::Map::setup();
-let proof = { 
+let proof = {
     // for recursion
     let k = ceil_log2(index.srs.g.len());
     let chals: Vec<_> = (0..k).map(|_| Fp::rand(rng)).collect();
@@ -77,12 +77,13 @@ The project is organized in the following way:
 
 * [book/](https://github.com/o1-labs/proof-systems/tree/master/book). The mina book, RFCs, and specifications.
 * [cairo/](https://github.com/o1-labs/proof-systems/tree/master/cairo). A Cairo runner written in rust.
-* [crypto/](https://github.com/o1-labs/proof-systems/tree/master/crypto). Interfaces for mina hashing and signature schemes.
 * [curves/](https://github.com/o1-labs/proof-systems/tree/master/curves). The elliptic curves we use (for now just the pasta curves).
 * [groupmap/](https://github.com/o1-labs/proof-systems/tree/master/groupmap). Used to convert elliptic curve elements to field elements.
+* [hasher/](https://github.com/o1-labs/proof-systems/tree/master/hasher). Interfaces for mina hashing.
 * [kimchi/](https://github.com/o1-labs/proof-systems/tree/master/kimchi). Our proof system.
 * [ocaml/](https://github.com/o1-labs/proof-systems/tree/master/ocaml). Ocaml bindings generator tool.
 * [oracle/](https://github.com/o1-labs/proof-systems/tree/master/oracle). Implementation of the poseidon hash function.
 * [poly-commitment/](https://github.com/o1-labs/proof-systems/tree/master/poly-commitment). Polynomial commitment code.
+* [signer/](https://github.com/o1-labs/proof-systems/tree/master/signer). Interfaces for mina signature schemes.
 * [tools/](https://github.com/o1-labs/proof-systems/tree/master/tools). Various tooling to help us work on kimchi.
 * [utils/](https://github.com/o1-labs/proof-systems/tree/master/utils). Collection of useful functions and traits.
