--- crypto/src/lib.rs
@@ -1,5 +1,0 @@
-#![deny(missing_docs)]
-#![doc = include_str!("../README.md")]
-
-pub mod hasher;
-pub mod signer;
